Fix autoloading gdb python extensions.
Centos package for 2.6 version of python-debuginfo provided a file
/usr/lib/debug/usr/lib64/libpython2.6.so.1.0.debug-gdb.py which cooperates
with gdb so that the commands like "py-bt" are automatically available.

This fix does the analogical thing but it uses
/usr/lib/debug/usr/bin/python2.7.debug-gdb.py as an entry point.

Note: to use debugging with gdb correctly, one still needs to change the spec
to include `--with-pydebug` option in the call to `./configure`. Without this
the frame information is optimized out and gdb is not that usefull.